requirements.txt
attrs==25.1.0
blessed==1.20.0
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.18.0
rich==13.9.3
six==1.16.0
wcwidth==0.2.13
